requirements.txt
sentence-transformers==2.2.2
faiss-cpu==1.8.0
google-generativeai==0.3.2
unidecode==1.3.7
numpy==1.26.4
flask==2.3.2
flask-cors==4.0.0
flask-sqlalchemy==3.0.5
flask-login==0.6.3
python-dotenv==1.0.1
werkzeug==2.3.7
PyPDF2==3.0.1
python-docx==0.8.11
huggingface-hub>=0.20.3,<0.26.0